src/ObservableShop/Product.Service/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-grpc==0.43b0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-instrumentation-psycopg2==0.43b0
opentelemetry-instrumentation-pika==0.43b0
psycopg2==2.9.9
pika==1.3.2